tests/wpt/web-platform-tests/css/motion/parsing/offset-distance-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Motion Path Module Level 1: parsing offset-distance with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/motion-1/#offset-distance-property">
<meta name="assert" content="offset-distance supports the full grammar '<length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("offset-distance", "10px");
test_valid_value("offset-distance", "20%");
test_valid_value("offset-distance", "calc(30px + 40%)");
test_valid_value("offset-distance", "0", "0px");
</script>
</body>
</html>
